пофиксил второе

+++ src/exercise2.c
@@ -2,7 +2,7 @@
 
 int main(int argc, char** argv){
 
-  unsigned  double che[10];
+        double che[10];
     for(int i = 0; i < 10; ++i){
         scanf("%lf", &che[i]);
     }
